Fix server side search (#1302)
* First working iteration

* Replace connected private route with own privateRoute component

* Replace PrivateRoute component in fhir-app
