+++ packages/react-utils/src/index.tsx
@@ -20,3 +20,4 @@ export * from './hooks/utils';
 export * from './components/PageHeader';
 export * from './components/ButtonLink';
 export * from './helpers/test-utils';
+export * from './components/PrivateRoute';
